Work with full representation also for realm roles
"composites" are not removed from role body anymore.
It seems this hack is not needed.
