+++ kcapi/rest/clients.py
@@ -1,6 +1,7 @@
 from copy import deepcopy
 
 from .crud import KeycloakCRUD
+from .roles import BaseRoleCRUD
 
 
 class Role():
@@ -23,21 +24,8 @@ class Role():
         return Composites(self.kc, self.value['id'])
 
 
-class ClientRoleCRUD(KeycloakCRUD):
-    _rest_params_read = {"briefRepresentation": False}
-
-    """
-    For RH SSO 7.4 (KC 9.0), role with attributes cannot be created.
-    We need to first create role, then update it to add attributes.
-
-    This will not be needed any more in RH SSO 7.5.
-    """
-    def create(self, payload):
-        ret = super().create(payload)
-        if "attributes" in payload:
-            role = self.findFirstByKV("name", payload["name"])
-            self.update(role["id"], payload).isOk()
-        return ret
+class ClientRoleCRUD(BaseRoleCRUD):
+    pass
 
 
 class Composites:
